authorization implemented

+++ src/api/note/handlers.js
@@ -1,7 +1,8 @@
 exports.NoteHandlers = (service, validator) => {
   const addNote = async (req, h) => {
     validator.validateNotePayload(req.payload)
-    const id = await service.addNote(req.payload)
+    const { id: credentialId } = req.auth.credentials
+    const id = await service.addNote(req.payload, credentialId)
 
     return h
       .response({
@@ -14,8 +15,9 @@ exports.NoteHandlers = (service, validator) => {
       .code(201)
   }
 
-  const getNotes = async () => {
-    const notes = await service.getNotes()
+  const getNotes = async (req) => {
+    const { id: credentialId } = req.auth.credentials
+    const notes = await service.getNotes(credentialId)
 
     return {
       status: 'success',
@@ -26,6 +28,8 @@ exports.NoteHandlers = (service, validator) => {
   }
 
   const getNoteById = async (req) => {
+    const { id: credentialId } = req.auth.credentials
+    await service.verifyNoteOwner(req.params.id, credentialId)
     const note = await service.getNoteById(req.params.id)
 
     return {
@@ -38,6 +42,8 @@ exports.NoteHandlers = (service, validator) => {
 
   const editNoteById = async (req) => {
     validator.validateNotePayload(req.payload)
+    const { id: credentialId } = req.auth.credentials
+    await service.verifyNoteOwner(req.params.id, credentialId)
     const note = await service.editNoteById(req.params.id, req.payload)
 
     return {
@@ -50,6 +56,8 @@ exports.NoteHandlers = (service, validator) => {
   }
 
   const deleteNoteById = async (req) => {
+    const { id: credentialId } = req.auth.credentials
+    await service.verifyNoteOwner(req.params.id, credentialId)
     await service.deleteNoteById(req.params.id)
 
     return {
